Move JavaFx related code into javafx package
